jmibeans-interact/jmibeans-shell-client/src/main/java/com/github/codeteapot/jmibeans/shell/client/secutity/auth/user/MachineShellPasswordName.java
package com.github.codeteapot.jmibeans.shell.client.secutity.auth.user;

import static java.util.Objects.requireNonNull;

import java.beans.ConstructorProperties;
import java.io.Serializable;

public class MachineShellPasswordName implements Serializable {

  private static final long serialVersionUID = 1L;

  private final String value;

  @ConstructorProperties({
      "value"
  })
  public MachineShellPasswordName(String value) {
    this.value = requireNonNull(value);
  }

  public String getValue() {
    return value;
  }

  @Override
  public int hashCode() {
    return value.hashCode();
  }

  @Override
  public boolean equals(Object obj) {
    if (obj == this) {
      return true;
    }
    if (obj instanceof MachineShellPasswordName) {
      MachineShellPasswordName passwordName = (MachineShellPasswordName) obj;
      return value.equals(passwordName.value);
    }
    return false;
  }

  @Override
  public String toString() {
    return value;
  }
}
